client minor refactor
